Move ctd-specific command utilities to ctd package
