+++ proxy/src/main/java/com/velocityctd/proxy/command/builtin/AlertRawCommand.java
@@ -21,12 +21,12 @@ import com.mojang.brigadier.Command;
 import com.mojang.brigadier.arguments.StringArgumentType;
 import com.mojang.brigadier.builder.LiteralArgumentBuilder;
 import com.mojang.brigadier.context.CommandContext;
+import com.velocityctd.proxy.command.CommandUtils;
 import com.velocityctd.proxy.redis.impl.packet.VelocityAlert;
 import com.velocitypowered.api.command.BrigadierCommand;
 import com.velocitypowered.api.command.CommandSource;
 import com.velocitypowered.api.permission.Tristate;
 import com.velocitypowered.proxy.VelocityServer;
-import com.velocitypowered.proxy.command.VelocityCommands;
 import com.velocitypowered.proxy.command.builtin.BuiltinCommand;
 import com.velocitypowered.proxy.util.ComponentUtils;
 import net.kyori.adventure.text.Component;
@@ -55,7 +55,7 @@ public class AlertRawCommand implements BuiltinCommand {
             .literalArgumentBuilder(label())
             .requires(source ->
                     source.getPermissionValue("velocity.command.alertraw") == Tristate.TRUE)
-            .executes(ctx -> VelocityCommands.emitUsage(ctx, label()))
+            .executes(ctx -> CommandUtils.emitUsage(ctx, label()))
             .then(BrigadierCommand
                     .requiredArgumentBuilder("message", StringArgumentType.greedyString())
                     .executes(this::alert));
