+++ proxy/src/main/java/com/velocityctd/proxy/command/builtin/FindCommand.java
@@ -22,13 +22,13 @@ import com.mojang.brigadier.arguments.StringArgumentType;
 import com.mojang.brigadier.builder.LiteralArgumentBuilder;
 import com.mojang.brigadier.builder.RequiredArgumentBuilder;
 import com.mojang.brigadier.context.CommandContext;
+import com.velocityctd.proxy.command.CommandUtils;
 import com.velocityctd.proxy.redis.VelocityRedis;
 import com.velocityctd.proxy.redis.impl.depot.PlayerEntry;
 import com.velocitypowered.api.command.BrigadierCommand;
 import com.velocitypowered.api.command.CommandSource;
 import com.velocitypowered.api.permission.Tristate;
 import com.velocitypowered.proxy.VelocityServer;
-import com.velocitypowered.proxy.command.VelocityCommands;
 import com.velocitypowered.proxy.command.builtin.BuiltinCommand;
 import com.velocitypowered.proxy.command.builtin.CommandMessages;
 import com.velocitypowered.proxy.connection.backend.VelocityServerConnection;
@@ -61,10 +61,10 @@ public class FindCommand implements BuiltinCommand {
             .literalArgumentBuilder(label())
             .requires(source ->
                     source.getPermissionValue("velocity.command.find") == Tristate.TRUE)
-            .executes(ctx -> VelocityCommands.emitUsage(ctx, label()));
+            .executes(ctx -> CommandUtils.emitUsage(ctx, label()));
     RequiredArgumentBuilder<CommandSource, String> playerNode = BrigadierCommand
             .requiredArgumentBuilder("player", StringArgumentType.word())
-            .suggests((ctx, builder) -> VelocityCommands.suggestPlayer(server, ctx, builder, true))
+            .suggests((ctx, builder) -> CommandUtils.suggestPlayer(server, ctx, builder, true))
             .executes(this::find);
 
     rootNode.then(playerNode);
