+++ proxy/src/main/java/com/velocityctd/proxy/command/builtin/PingCommand.java
@@ -21,12 +21,12 @@ import com.mojang.brigadier.Command;
 import com.mojang.brigadier.arguments.StringArgumentType;
 import com.mojang.brigadier.builder.LiteralArgumentBuilder;
 import com.mojang.brigadier.context.CommandContext;
+import com.velocityctd.proxy.command.CommandUtils;
 import com.velocityctd.proxy.redis.impl.transaction.VelocityGetPlayerPing;
 import com.velocitypowered.api.command.BrigadierCommand;
 import com.velocitypowered.api.command.CommandSource;
 import com.velocitypowered.api.permission.Tristate;
 import com.velocitypowered.proxy.VelocityServer;
-import com.velocitypowered.proxy.command.VelocityCommands;
 import com.velocitypowered.proxy.command.builtin.BuiltinCommand;
 import com.velocitypowered.proxy.command.builtin.CommandMessages;
 import com.velocitypowered.proxy.connection.client.ConnectedPlayer;
@@ -53,24 +53,24 @@ public class PingCommand implements BuiltinCommand {
   @Override
   public BrigadierCommand build() {
     LiteralArgumentBuilder<CommandSource> node = BrigadierCommand.literalArgumentBuilder(label())
-            .requires(source -> source.getPermissionValue("velocity.command.ping") == Tristate.TRUE)
-            .then(
-                    BrigadierCommand.requiredArgumentBuilder("player", StringArgumentType.word())
-                            .requires(source -> source.getPermissionValue("velocity.command.ping.others") == Tristate.TRUE)
-                            .suggests((ctx, builder) -> VelocityCommands.suggestPlayer(server, ctx, builder, true))
-                            .executes(context -> {
-                              String player = context.getArgument("player", String.class);
-                              return this.getPing(context, player);
-                            })
-            )
-            .executes(context -> {
-              if (context.getSource() instanceof ConnectedPlayer player) {
-                return this.getPing(context, player.getUsername());
-              } else {
-                context.getSource().sendMessage(CommandMessages.PLAYERS_ONLY);
-                return 0;
-              }
-            });
+        .requires(source -> source.getPermissionValue("velocity.command.ping") == Tristate.TRUE)
+        .then(
+            BrigadierCommand.requiredArgumentBuilder("player", StringArgumentType.word())
+                .requires(source -> source.getPermissionValue("velocity.command.ping.others") == Tristate.TRUE)
+                .suggests((ctx, builder) -> CommandUtils.suggestPlayer(server, ctx, builder, true))
+                .executes(context -> {
+                  String player = context.getArgument("player", String.class);
+                  return this.getPing(context, player);
+                })
+        )
+        .executes(context -> {
+          if (context.getSource() instanceof ConnectedPlayer player) {
+            return this.getPing(context, player.getUsername());
+          } else {
+            context.getSource().sendMessage(CommandMessages.PLAYERS_ONLY);
+            return 0;
+          }
+        });
 
     return new BrigadierCommand(node);
   }
@@ -90,36 +90,36 @@ public class PingCommand implements BuiltinCommand {
 
       if (ping == -1L) {
         context.getSource().sendMessage(
-                Component.translatable("velocity.command.ping.unknown", NamedTextColor.RED)
-                        .arguments(Argument.string("player", player.getUsername())));
+            Component.translatable("velocity.command.ping.unknown", NamedTextColor.RED)
+                .arguments(Argument.string("player", player.getUsername())));
         return 0;
       }
 
       context.getSource().sendMessage(
-              Component.translatable("velocity.command.ping.self", NamedTextColor.GREEN)
-                      .arguments(Argument.numeric("ping", ping))
+          Component.translatable("velocity.command.ping.self", NamedTextColor.GREEN)
+              .arguments(Argument.numeric("ping", ping))
       );
     } else {
       if (server.isRedisEnabled()) {
         if (!this.server.getRedis().getPlayerService().isPlayerOnline(username)) {
           context.getSource().sendMessage(Component.translatable("velocity.command.player-not-found")
-                  .arguments(Argument.string("player", username)));
+              .arguments(Argument.string("player", username)));
           return -1;
         }
 
         new VelocityGetPlayerPing(context.getSource(), username)
-                .publish();
+            .publish();
       } else {
         if (player == null) {
           context.getSource().sendMessage(Component.translatable("velocity.command.player-not-found")
-                  .arguments(Argument.string("player", username)));
+              .arguments(Argument.string("player", username)));
           return -1;
         }
 
         Component component = Component.translatable("velocity.command.ping.other", NamedTextColor.GREEN)
-                .arguments(
-                        Argument.string("player", player.getUsername()),
-                        Argument.numeric("ping", player.getPing()));
+            .arguments(
+                Argument.string("player", player.getUsername()),
+                Argument.numeric("ping", player.getPing()));
 
         context.getSource().sendMessage(component);
       }
