+++ proxy/src/main/java/com/velocitypowered/proxy/command/builtin/ShutdownCommand.java
@@ -21,11 +21,11 @@ import com.mojang.brigadier.Command;
 import com.mojang.brigadier.arguments.StringArgumentType;
 import com.mojang.brigadier.builder.LiteralArgumentBuilder;
 import com.mojang.brigadier.builder.RequiredArgumentBuilder;
+import com.velocityctd.proxy.command.CommandUtils;
 import com.velocitypowered.api.command.BrigadierCommand;
 import com.velocitypowered.api.command.CommandSource;
 import com.velocitypowered.api.proxy.ConsoleCommandSource;
 import com.velocitypowered.proxy.VelocityServer;
-import com.velocitypowered.proxy.command.VelocityCommands;
 import java.util.List;
 import org.checkerframework.checker.nullness.qual.NonNull;
 
@@ -62,7 +62,7 @@ public class ShutdownCommand implements BuiltinCommand {
                             StringArgumentType.greedyString())
                     .executes(context -> {
                       String reason = context.getArgument("reason", String.class);
-                      server.shutdown(true, VelocityCommands.deserializeComponent(reason));
+                      server.shutdown(true, CommandUtils.deserializeComponent(reason));
                       return Command.SINGLE_SUCCESS;
                     })
             ).build());
